Move ZetaTestHelper.hpp to own file
